refactor: split user and subscription services
